apps/api/pyproject.toml
[project]
name = "sky-scanner-api"
version = "0.1.0"
description = "FastAPI backend for flight search"
requires-python = ">=3.13"
dependencies = [
    "sky-scanner-core",
    "sky-scanner-db",
    "sky-scanner-ml",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "redis>=5.2.0",
    "pydantic-settings>=2.7.0",
    "celery>=5.4.0",
    "pyjwt[crypto]>=2.10.0",
    "bcrypt>=4.0.0",
    "pydantic[email]>=2.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sky_scanner_api"]

[tool.uv.sources]
sky-scanner-core = { workspace = true }
sky-scanner-db = { workspace = true }
sky-scanner-ml = { workspace = true }
